requirements.txt
streamlit
pandas
plotly
python-docx
reportlab
openpyxl
